% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_GE_helperfunctions.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Internal: Dataframe Checker}
\usage{
check_dataframe(x)
}
\description{
Check whether the insertet object is a dataframe or not.
}
\keyword{internal}
