% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{describe}
\alias{describe}
\title{Describe continuous variables, factors and logicals}
\usage{
describe(x, num.desc = c("mean", "median", "var", "sd", "valid.n"),
  xname = NA, horizontal = FALSE)
}
\arguments{
\item{x}{}

\item{num.desc}{}

\item{xname}{}

\item{horizontal}{}
}
\description{
Describe continuous variables, factors and logicals
}
