% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_suggest_fixes.R
\name{display_fix_suggestions}
\alias{display_fix_suggestions}
\title{Display fix suggestions in a readable format}
\usage{
display_fix_suggestions(fixes, copy_to_clipboard = FALSE)
}
\description{
Display fix suggestions in a readable format
}
\keyword{internal}
