% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGenemy.R
\name{GGenemy}
\alias{GGenemy}
\title{GGenemy}
\usage{
GGenemy()
}
\value{
A Shiny App.
}
\description{
Starts the Shiny App in which a dataset can be reviewed, several summary statistics can be
calculated and conditional plots can be drawn.
}
\examples{
# coming soon.
}
