% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_densities.R
\name{plot_conditional_densities}
\alias{plot_conditional_densities}
\title{Plot conditional densities}
\usage{
plot_conditional_densities(var_name, dataset, n_quantiles = 5)
}
\arguments{
\item{var_name}{A variable from your \code{dataset}.}

\item{dataset}{A dataframe.}

\item{n_quantiles}{Number of quantiles you want to partition \code{var_name} into.}
}
\value{

}
\description{
Plot densities of continuous or categorical variables conditioned on quantiles of continuous
or categories of categorical variables.
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a){
  if (runif(1) < 0.5) rnorm(n)
    else rgamma(n, 0.5)
}))

plot_conditional_densities("V1", variables)     
}
