% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GGenemy.R
\name{plot_GGenemy}
\alias{plot_GGenemy}
\title{Plot GGenemy}
\usage{
plot_GGenemy(dataset, given_var, var_to_plot = NULL, n_quantiles = 5,
  boxplot = FALSE, selfquantiles = NULL, remaining = TRUE)
}
\arguments{
\item{dataset}{A dataframe.}

\item{given_var}{A variable from your \code{dataset}.}

\item{var_to_plot}{The variables to plot. When not selecting any variable, all variables from your \code{dataset}
will be plotted.}

\item{n_quantiles}{Number of quantiles you want to partition \code{given_var}
into.}

\item{boxplot}{logical. If TRUE boxplots will be presented instead of densities
Also possible to use a vector with the names of variables, which should be plotted as boxplots instead of the density.}

\item{selfquantiles}{Vector of n times two Values or matrix with 2 columns.
The first value as min and the second value as max of the self selected quantile.}

\item{remaining}{logical. If TRUE the remaining values not within selfquantiles will be plotted to
category remaining.}
}
\value{
Several ggplot graphics. The total number of plots will equal the
  number of variables in your \code{dataset} or selected variables from \code{var_to_plot}
}
\description{
Plots densities of continuous variables, conditioned on quantiles of
continuous variables or categories of categorical variables. For categorical
variables, boxplots or barplots will be created, depending on whether the
variable to be conditioned on is continuous or categorical.
Either the user can use a number of same sized quantiles or
can select the size of own quantiles
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}))


plot_GGenemy(variables, "V1")
}
