% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_conditional_densities.R
\name{plot_multiple_conditional_densities}
\alias{plot_multiple_conditional_densities}
\title{Plot multiple conditional densities}
\usage{
plot_multiple_conditional_densities(dataset, var_name, n_quantiles = 5,
  var_to_cond_on)
}
\arguments{
\item{dataset}{A dataframe.}

\item{var_name}{A variable from your \code{dataset}.}

\item{n_quantiles}{Number of quantiles you want to partition \code{var_name} into.}

\item{var_to_cond_on}{Vector of Variables to condition on from your \code{dataset}.}
}
\value{
Several ggplot graphics. The total number of plots will equal the
  number of variables in your \code{dataset}.
}
\description{
Plots densities of chosen continuous variables, conditioned on quantiles of
continuous variables or categories of categorical variables. For categorical
variables, boxplots or barplots will be created, depending on whether the
variable to be conditioned on is continuous or categorical.
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}))

plot_multiple_conditional_densities(variables, "V1", n_quantiles = 5, c("V2","V3"))
}
