% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_selected_quantile_densities.R
\name{plot_multiple_selected_quantile_densities}
\alias{plot_multiple_selected_quantile_densities}
\title{Plot multiple selected quantile densities}
\usage{
plot_multiple_selected_quantile_densities(dataset, given_var,
  firstquant = NA, secondquant = NA, thirdquant = NA, var_to_plot)
}
\arguments{
\item{dataset}{A dataframe.}

\item{given_var}{A variable from your \code{dataset}.}

\item{firstquant}{Vector of two Values, with min and max value of the first self selected quantile}

\item{secondquant}{Vector of two Values, with min and max value of the second self selected quantile}

\item{thirdquant}{Vector of two Values, with min and max value of the third self selected quantile}

\item{var_to_plot}{A Vector of Variables to condition on from your \code{dataset}.}
}
\value{
Several ggplot graphics. The total number of plots will equal the 
number of chosen variables of your \code{dataset}.
}
\description{
Plots densities of chosen continuous variables, conditioned on self selected quantiles of
continuous variables. For categorical
variables, boxplots will be created.
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}))

plot_multiple_selected_quantile_densities(variables, "V1", c(0,1),c(2,3),
var_to_plot = c("V2","V3"))
}
