% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_conditional_density.R
\name{plot_single_conditional_density}
\alias{plot_single_conditional_density}
\title{Plot a single conditional density}
\usage{
plot_single_conditional_density(dataset, var_name, n_quantiles = 5,
  var_to_cond_on)
}
\arguments{
\item{dataset}{A dataframe.}

\item{var_name}{The variable of interest from your \code{dataset}.}

\item{n_quantiles}{Number of quantiles you want to partition \code{var_name} into.}

\item{var_to_cond_on}{A variable to condition on from your \code{dataset}.}
}
\value{
A ggplot graphic.
}
\description{
Plots the conditional density of a continuous variable conditioned on quantiles of
another continuous variable or categories of a categorical variable. Other plot types for other classes of variables.
}
\details{
If the variable of interest is continuous, the function will create 
a conditional density plot. Depending on whether the variable to be conditioned
on is continuous itself or a factor, the dependent variable will be
conditioned on its quantiles or categories.
If the variable of interest is categorical, a boxplot or barplot will be created,
depending on whether the conditional variable is continuous or categorical.
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}))

plot_single_conditional_density(variables, "V1", 5, "V2")
}
