% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_selected_quantile_density.R
\name{plot_single_selected_quantile_density}
\alias{plot_single_selected_quantile_density}
\title{Plot multiple conditional densities}
\usage{
plot_single_selected_quantile_density(dataset, given_var,
  firstquant = NA, secondquant = NA, thirdquant = NA, var_to_plot)
}
\arguments{
\item{dataset}{A dataframe.}

\item{given_var}{A variable from your \code{dataset}.}

\item{firstquant}{Vector of two Values, with min and max value of the first self selected quantile}

\item{secondquant}{Vector of two Values, with min and max value of the second self selected quantile}

\item{thirdquant}{Vector of two Values, with min and max value of the third self selected quantile}

\item{var_to_plot}{A Variable to condition on from your \code{dataset}.}
}
\value{
A ggplot graphic
}
\description{
Plots a densitiy of a chosen continuous variable, conditioned on self selected quantiles of
continuous variables. For categorical
variables, boxplots will be created.
}
\examples{
n <- 500
number_variables <- 5

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}
))

plot_single_selected_quantile_density(variables, "V1", c(0,1), c(1.2,1.8),
c(2,3), var_to_plot = "V2")
}
