% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sum_stats.R
\name{plot_sum_stats}
\alias{plot_sum_stats}
\title{Plot Summary Statistics}
\usage{
plot_sum_stats(dataset, given_var, stats = 1, n_quantiles = 5)
}
\arguments{
\item{dataset}{A data frame. Factors and logicals will be removed.}

\item{given_var}{A variable from your \code{dataset} which you want to set as a condition.}

\item{stats}{Vector of summary statistics you want to plot.}

\item{n_quantiles}{Number of quantiles you want to partition \code{given_var}
into, with a maximum of 10.}
}
\value{
One plot for each statistic, showing all variables conditioned on
the given variable.
}
\description{
Calculates the conditional mean, variance, skewness and kurtosis for continuous variables
based on a range of quantiles of a given continuous variable and displays them
as line plots.
}
\details{
The function will return one plot for each of the chosen summary statistics. The
values will be displayed as line plots, each point representing the value of
the statistic conditioned on a specific quantile of the given variable.
}
\examples{
data(iris)
iris <- iris[, -5]
plot_sum_stats(iris, "Sepal.Length", stats = c(1, 2, 3, 4), n_quantiles = 5)
}
