% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_suggest_fixes.R
\name{print.gg_fix_suggestions}
\alias{print.gg_fix_suggestions}
\title{Print method for fix suggestions}
\usage{
\method{print}{gg_fix_suggestions}(x, ...)
}
\arguments{
\item{x}{A gg_fix_suggestions object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for fix suggestions
}
