% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{print_sum_stats}
\alias{print_sum_stats}
\title{Print results of \code{sum_stats()} nicely}
\usage{
print_sum_stats(x, given_var)
}
\arguments{
\item{x}{An object created by sum_stats().}

\item{given_var}{The given variable you used to create \code{x}. The variable's
name is given as a string.}
}
\value{
The same results as in \code{sum_stats()}, but printed nicely.
}
\description{
Use this function to print your results from \code{sum_stats()} in a more
reader-friendly way.
}
\examples{
data(iris)
x <- sum_stats(iris, "Sepal.Length", stats = c("mean", "var"), n_quantiles = 5) 
print_sum_stats(x, "Sepal.Length")
}
