% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_GE_helperfunctions.R
\name{search_dataframe}
\alias{search_dataframe}
\title{Internal: Seatch Dataframe in GE}
\usage{
search_dataframe()
}
\description{
Searches for Dataframes in the Global Environment.
}
\keyword{internal}
