% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Title}
\usage{
sum_stats(var_name, dataset, n_quantiles = 5, n_sum_stats = 3)
}
\arguments{
\item{var_name}{A variable from your \code{dataset}.}

\item{dataset}{A dataframe.}

\item{n_quantiles}{Number of quantiles you want to partition \code{var_name} into.}

\item{n_sum_stats}{Number of summary statistics you want to print. Order of the 
statistics is: conditional mean, conditional variance, conditional skewness, 
conditional kurtosis.}
}
\value{
A list with one entry for each summary statistic.
}
\description{
Title
}
