% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunction.R
\name{testfunction}
\alias{testfunction}
\title{Calculate the sum of all vector elements}
\usage{
testfunction(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
Calculate the sum of all vector elements
}
\examples{
testfunction(c(1,2,3,4,5))
}
