% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_GE_helperfunctions.R
\name{search_dataframe}
\alias{search_dataframe}
\title{Searcher GE Dataframes}
\usage{
search_dataframe()
}
\arguments{
\item{x}{Element of Global Environment}
}
\value{
The names of dataframes from the Global Environment.
}
\description{
Helpfunctions to search for dataframes in Global Environment, while using Shiny
}
