% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Summary Statistics}
\usage{
sum_stats(var_name, dataset, n_quantiles = 5, n_sum_stats = 3)
}
\arguments{
\item{var_name}{A variable from your \code{dataset}.}

\item{dataset}{A dataframe.}

\item{n_quantiles}{Number of quantiles you want to partition \code{var_name}
into.}

\item{n_sum_stats}{Number of summary statistics you want to print. The order
of the statistics is as follows: conditional mean, conditional variance,
conditional skewness, conditional kurtosis.}
}
\value{
A list with one entry for each summary statistic.
}
\description{
Calculates the means, variances, skewness and kurtosis for a range of
quantiles of continuous variables.
}
\examples{
n <- 500
number_variables <- 5
n_quantiles <- 4

set.seed(42)
variables <- as.data.frame(sapply(1:number_variables, function(a) {
  if (runif(1) < 0.5) {
    rnorm(n)
  } else {
    rgamma(n, 0.5)
  }
}))

sum_stats("V1", variables, n_quantiles, n_sum_stats = 4)
}
